fix integration tests
