+++ tests/integ_tests/inference/native_inference/test_open_clip_inference_pipeline.py
@@ -3,9 +3,13 @@ import numpy as np
 from integ_tests.inference.inference_test_case import InferenceTestCase
 from integ_tests.marqo_test import TestImageUrls
 from marqo.core.inference.api import *
+from marqo.core.inference.device_manager import DeviceManager
 from marqo.inference.native_inference.local_inference import NativeInferenceLocal
 
 class TestOpenCLIPInferencePipeline(InferenceTestCase):
+    
+    def setUp(self):
+        self.inference = NativeInferenceLocal(device_manager=DeviceManager())
 
     def test_inference_text_no_chunk_no_prefix(self):
         """Test that the pipeline returns the embeddings for the two texts without chunking or prefix."""
@@ -27,7 +31,7 @@ class TestOpenCLIPInferencePipeline(InferenceTestCase):
             )
         )
 
-        results = NativeInferenceLocal().vectorise(text_inference_request)
+        results = self.inference.vectorise(text_inference_request)
 
         self.assertTrue(isinstance(results, InferenceResult))
         self.assertTrue(isinstance(results.result, list))
@@ -76,7 +80,7 @@ class TestOpenCLIPInferencePipeline(InferenceTestCase):
             )
         )
 
-        results = NativeInferenceLocal().vectorise(image_inference_request)
+        results = self.inference.vectorise(image_inference_request)
 
         self.assertTrue(isinstance(results, InferenceResult))
         self.assertTrue(isinstance(results.result, list))
@@ -126,7 +130,7 @@ class TestOpenCLIPInferencePipeline(InferenceTestCase):
             )
         )
 
-        results = NativeInferenceLocal().vectorise(image_inference_request)
+        results = self.inference.vectorise(image_inference_request)
 
         self.assertTrue(isinstance(results, InferenceResult))
         self.assertTrue(isinstance(results.result, list))
